+++ Changelog.md
@@ -4,6 +4,8 @@
 
 ## Next release
 
+- New statements added: <start transaction statement>,
+  <commit statement>, <rollback statement>.
 - Refactoring: introducing non terminal symbols from SQL grammar.
 - eesql:query_specification/0 renamed to eesql:query_spec/0